+++ solutions/ColumnFormatter/src/webparts/columnFormatter/components/ColumnFormatterWelcome.tsx
@@ -93,7 +93,7 @@ class ColumnFormatterWelcome_ extends React.Component<IColumnFormatterWelcomePro
 
     //TEMP (helpful for testing and skipping the welcome altogether)
     //this.props.launchEditor(undefined,columnTypes.text);
-    //this.props.launchEditor('Data Bars', columnTypes.number);
+    //this.props.launchEditor('Donut', columnTypes.number);
     //this.props.launchEditor('Severity', columnTypes.text);
     //this.props.launchEditor('Start Flow', columnTypes.text);